[logs+apm] Add options to allow resetting connections at configurable interval (#5718)
* [logs] Add option to reset HTTP connections at configured interval

Configurable with `logs_config.connection_reset_interval`.

Disabled by default (for now).

* [apm] Reset outgoing HTTP connections at configurable interval

Disabled by default.

Logic shared with Logs Agent.
