+++ pkg/logs/config/endpoints.go
@@ -11,13 +11,14 @@ import (
 
 // Endpoint holds all the organization and network parameters to send logs to Datadog.
 type Endpoint struct {
-	APIKey           string `mapstructure:"api_key" json:"api_key"`
-	Host             string
-	Port             int
-	UseSSL           bool
-	UseCompression   bool `mapstructure:"use_compression" json:"use_compression"`
-	CompressionLevel int  `mapstructure:"compression_level" json:"compression_level"`
-	ProxyAddress     string
+	APIKey                  string `mapstructure:"api_key" json:"api_key"`
+	Host                    string
+	Port                    int
+	UseSSL                  bool
+	UseCompression          bool `mapstructure:"use_compression" json:"use_compression"`
+	CompressionLevel        int  `mapstructure:"compression_level" json:"compression_level"`
+	ProxyAddress            string
+	ConnectionResetInterval time.Duration
 }
 
 // Endpoints holds the main endpoint and additional ones to dualship logs.
