+++ pkg/trace/config/config_test.go
@@ -10,6 +10,7 @@ import (
 	"regexp"
 	"strings"
 	"testing"
+	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/stretchr/testify/assert"
@@ -235,6 +236,7 @@ func TestUndocumentedYamlConfig(t *testing.T) {
 	assert.Equal(100.0, c.MaxTPS)
 	assert.Equal(1000.0, c.MaxEPS)
 	assert.Equal(25, c.ReceiverPort)
+	assert.Equal(120*time.Second, c.ConnectionResetInterval)
 	// watchdog
 	assert.Equal(0.07, c.MaxCPU)
 	assert.Equal(30e6, c.MaxMemory)
