+++ pkg/util/http/client.go
@@ -6,15 +6,11 @@
 package http
 
 import (
-	"crypto/tls"
-	"fmt"
-	"net"
 	"net/http"
-	"net/url"
 	"regexp"
+	"sync"
 	"time"
 
-	"github.com/DataDog/datadog-agent/pkg/config"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -31,86 +27,58 @@ func SanitizeURL(message string) string {
 	return apiKeyRegExp.ReplaceAllString(message, apiKeyReplacement)
 }
 
-// CreateHTTPTransport creates an *http.Transport for use in the agent
-func CreateHTTPTransport() *http.Transport {
-	tlsConfig := &tls.Config{
-		InsecureSkipVerify: config.Datadog.GetBool("skip_ssl_validation"),
-	}
+// ResetClient wraps (http.Client).Do and resets the underlying connections at the
+// configured interval
+type ResetClient struct {
+	httpClientFactory func() *http.Client
+	resetInterval     time.Duration
 
-	if config.Datadog.GetBool("force_tls_12") {
-		tlsConfig.MinVersion = tls.VersionTLS12
-	}
-
-	// Most of the following timeouts are a copy of Golang http.DefaultTransport
-	// They are mostly used to act as safeguards in case we forget to add a general
-	// timeout to our http clients.
-	transport := &http.Transport{
-		TLSClientConfig: tlsConfig,
-		DialContext: (&net.Dialer{
-			Timeout: 30 * time.Second,
-			// Enables TCP keepalives to detect broken connections
-			KeepAlive: 30 * time.Second,
-			// Disable RFC 6555 Fast Fallback ("Happy Eyeballs")
-			FallbackDelay: -1 * time.Nanosecond,
-		}).DialContext,
-		MaxIdleConns:        100,
-		MaxIdleConnsPerHost: 5,
-		// This parameter is set to avoid connections sitting idle in the pool indefinitely
-		IdleConnTimeout:       90 * time.Second,
-		TLSHandshakeTimeout:   10 * time.Second,
-		ExpectContinueTimeout: 1 * time.Second,
-	}
-
-	if proxies := config.GetProxies(); proxies != nil {
-		transport.Proxy = GetProxyTransportFunc(proxies)
-	}
-
-	return transport
+	mu         sync.RWMutex
+	httpClient *http.Client
+	lastReset  time.Time
 }
 
-// GetProxyTransportFunc return a proxy function for a http.Transport that
-// would return the right proxy depending on the configuration.
-func GetProxyTransportFunc(p *config.Proxy) func(*http.Request) (*url.URL, error) {
-	return func(r *http.Request) (*url.URL, error) {
-		// check no_proxy list first
-		for _, host := range p.NoProxy {
-			if r.URL.Host == host {
-				log.Debugf("URL match no_proxy list item '%s': not using any proxy", host)
-				return nil, nil
-			}
-		}
-
-		// check proxy by scheme
-		confProxy := ""
-		if r.URL.Scheme == "http" {
-			confProxy = p.HTTP
-		} else if r.URL.Scheme == "https" {
-			confProxy = p.HTTPS
-		} else {
-			log.Warnf("Proxy configuration do not support scheme '%s'", r.URL.Scheme)
-		}
-
-		if confProxy != "" {
-			proxyURL, err := url.Parse(confProxy)
-			if err != nil {
-				err := fmt.Errorf("Could not parse the proxy URL for scheme %s from configuration: %s", r.URL.Scheme, err)
-				log.Error(err.Error())
-				return nil, err
-			}
-			userInfo := ""
-			if proxyURL.User != nil {
-				if _, isSet := proxyURL.User.Password(); isSet {
-					userInfo = "*****:*****@"
-				} else {
-					userInfo = "*****@"
-				}
-			}
-
-			log.Debugf("Using proxy %s://%s%s for URL '%s'", proxyURL.Scheme, userInfo, proxyURL.Host, SanitizeURL(r.URL.String()))
-			return proxyURL, nil
-		}
-
-		// no proxy set for this request
-		return nil, nil
+// NewResetClient returns an initialized Client resetting connections at the passed resetInterval ("0"
+// means that no reset is performed).
+// The underlying http.Client used will be created using the passed http client factory.
+func NewResetClient(resetInterval time.Duration, httpClientFactory func() *http.Client) *ResetClient {
+	return &ResetClient{
+		httpClientFactory: httpClientFactory,
+		resetInterval:     resetInterval,
+		httpClient:        httpClientFactory(),
+		lastReset:         time.Now(),
 	}
+}
+
+// Do wraps (http.Client).Do. Thread safe.
+func (c *ResetClient) Do(req *http.Request) (*http.Response, error) {
+	c.checkReset()
+
+	c.mu.RLock()
+	httpClient := c.httpClient
+	c.mu.RUnlock()
+
+	return httpClient.Do(req)
+}
+
+// checkReset checks whether a client reset should be performed, and performs it
+// if so
+func (c *ResetClient) checkReset() {
+	if c.resetInterval == 0 {
+		return
+	}
+
+	c.mu.Lock()
+	defer c.mu.Unlock()
+	if time.Since(c.lastReset) < c.resetInterval {
+		return
+	}
+
+	log.Debug("Resetting HTTP client's connections")
+	c.lastReset = time.Now()
+	// Close idle connections on underlying client. Safe to do while other goroutines use the client.
+	// This is a best effort: if other goroutine(s) are currently using the client,
+	// the related open connection(s) will remain open until the client is GC'ed
+	c.httpClient.CloseIdleConnections()
+	c.httpClient = c.httpClientFactory()
 }
